Add Learn more component and use it for blueprint and CLI
